Add resources for X-Ray 1.5.03

+++ gamedata_cs/shaders/r3/shared/wmark.h
@@ -7,7 +7,7 @@
 #define	MAX_SHIFT		0.011f
 #define RANGE			100.f
 
-uniform half3		eye_direction;
+uniform float3		eye_direction;
 
 float4 	wmark_shift 	(float3 pos, float3 norm)
 {
